Fix more CUDA compile errors

+++ gpu/src/cuda/shaders/utils.cuh
@@ -1,4 +1,4 @@
 #pragma once
 
 /// Reverses the `log2(size)` first bits of `i`
-__device__ uint reverse_index(uint i, usize size);
+__device__ uint reverse_index(uint i, uint size);
